usize -> u64

+++ benches/bench.rs
@@ -6,7 +6,7 @@ use criterion::{criterion_group, criterion_main, Criterion};
 
 use buffered_offset_reader::*;
 
-fn make_temp_file(chunk_size: u8, chunk_count: usize) -> (File, Vec<u8>) {
+fn make_temp_file(chunk_size: u8, chunk_count: u64) -> (File, Vec<u8>) {
     let mut file = tempfile().unwrap();
     let chunk = (0..chunk_size).into_iter().collect::<Vec<u8>>();
 
@@ -17,7 +17,7 @@ fn make_temp_file(chunk_size: u8, chunk_count: usize) -> (File, Vec<u8>) {
 }
 
 const CHUNK_SIZE: u8 = 64;
-const CHUNK_COUNT: usize = 1024;
+const CHUNK_COUNT: u64 = 1024;
 
 fn read_file(c: &mut Criterion) {
     let (file, chunk) = make_temp_file(CHUNK_SIZE, CHUNK_COUNT);
@@ -26,7 +26,7 @@ fn read_file(c: &mut Criterion) {
         let mut tmp = vec![0; CHUNK_SIZE as usize];
         b.iter(|| {
             for i in 0..CHUNK_COUNT {
-                file.read_at(&mut tmp, i * CHUNK_SIZE as usize).unwrap();
+                file.read_at(&mut tmp, i * CHUNK_SIZE as u64).unwrap();
                 assert_eq!(&tmp, &chunk);
             }
         });
@@ -42,7 +42,7 @@ fn read_buffered(c: &mut Criterion) {
         let mut tmp = vec![0; CHUNK_SIZE as usize];
         b.iter(|| {
             for i in 0..CHUNK_COUNT {
-                r.read_at(&mut tmp, i * CHUNK_SIZE as usize).unwrap();
+                r.read_at(&mut tmp, i * CHUNK_SIZE as u64).unwrap();
                 assert_eq!(&tmp, &chunk);
             }
         });
